Patch 1.0.3
Código relacionado à conexão com o banco de dados SQL foi jogado para um arquivo próprio.
