+++ Itens.py
@@ -2,15 +2,15 @@
 #Não é necessário importar nada
 
 class Item:
-    def __init__(self, nome, qtd): #Construtor(Inicializador) da classe item
+    def __init__(self, nome, quantidade): #Construtor(Inicializador) da classe item
         self.nome = nome
-        self.qtd = qtd
+        self.quantidade = quantidade
 
     def to_dict(self): #Criação do dicionário
-        return {"tipo": self.__class__.__name__, "nome": self.nome, "qtd": self.qtd}
+        return {"tipo": self.__class__.__name__, "nome": self.nome, "quantidade": self.quantidade}
 
     @classmethod #médoto de classe
-    def from_dict(cls, data):
+    def from_dict(classe, data):
         #Decide o tipo de item que é criado com base no que é pedido
         tipo = data.pop("tipo")
         if tipo == "Medicacao":
@@ -21,8 +21,8 @@ class Item:
             return ProdutoLimpeza(**data)
 #Medicação é uma subclasse de item que só pode ser acesada por farmacêuticos
 class Medicacao(Item):
-    def __init__(self, nome,  qtd):
-        super().__init__(nome,  qtd)
+    def __init__(self, nome,  quantidade):
+        super().__init__(nome,  quantidade)
 
     def to_dict(self):
         data = super().to_dict()
@@ -30,8 +30,8 @@ class Medicacao(Item):
 
 #Equipamento Medico é uma subclasse de item que só pode ser acessada por enfermeiros
 class EquipamentoMedico(Item):
-    def __init__(self, nome,  qtd):
-        super().__init__(nome,  qtd)
+    def __init__(self, nome,  quantidade):
+        super().__init__(nome,  quantidade)
 
 
     def to_dict(self):
@@ -40,8 +40,8 @@ class EquipamentoMedico(Item):
 
 #Produto de Limpeza é uma subclasse de item que só pode ser acessada por Auxiliares de Serviços Gerais
 class ProdutoLimpeza(Item):
-    def __init__(self, nome, qtd):
-        super().__init__(nome,  qtd)
+    def __init__(self, nome, quantidade):
+        super().__init__(nome,  quantidade)
 
 
     def to_dict(self):
